Incorporated OOP into commands - Deploy-commands is now activated on arg - Removed some not very common functions from Bot interface - Everything for commands moved inside class
